refactor: try-catch, doc, comment, helper/saveFiles

+++ controllers/editPictureInStorage.js
@@ -1,3 +1,8 @@
+/**
+ * @description     - user edit a posted picture. This function save new picture to storage
+ * @routes          - [PUT] /post
+ */
+
 const multer = require("multer");
 const fs = require('fs');
 const path = require('path');
@@ -6,19 +11,31 @@ const readPictureByPictureIds = require('../models/readPictureByPictureIds');
 // delete old picture file and save new picture
 const storage = multer.diskStorage({
     destination: async (req, file, cb) => {
-        picture = await readPictureByPictureIds(req.body.picture_id);
-        req.body.picture = picture[0];
-        req.body.originalfilename = file.originalname;
-        fs.unlinkSync(path.join(__dirname,'..',`/storage/${req.body.user_id}/img/${req.body.picture.uuid}.${req.body.picture.filetype}`));
-        const dest = path.join(__dirname,'..',`/storage/${req.body.user_id}/img`);
-        cb(null, dest);
-        //cb(null, path.join(__dirname,'..','storage/test')); // for debug
+        try{
+            const picture = await readPictureByPictureIds(req.body.picture_id);
+            console.log(picture);
+            let isValidUpdate = false;
+            console.log(picture[0].user_id);
+            if((picture.length === 1) && (parseInt(req.body.user_id) === picture[0].user_id)) 
+                isValidUpdate = true;
+            console.log(isValidUpdate);
+            if(isValidUpdate) {
+                req.body.picture = picture[0];
+                req.body.originalfilename = file.originalname;
+                fs.unlinkSync(path.join(__dirname,'..',`/storage/${req.body.user_id}/img/${req.body.picture.uuid}.${req.body.picture.filetype}`));
+                const dest = path.join(__dirname,'..',`/storage/${req.body.user_id}/img`);
+                cb(null, dest);
+            }
+        } catch (error) {
+            console.error(error);
+        }
     },
     filename: async (req, file, cb) => {
-        x = file.mimetype.split("/");
-        filetype = x[x.length - 1];
-        cb(null, `${req.body.picture.uuid}.${filetype}`);
-        // cb(null, `${require('uuid').v4()}.${filetype}`); // for debug
+        if(isValidUpdate) {
+            x = file.mimetype.split("/");
+            filetype = x[x.length - 1];
+            cb(null, `${req.body.picture.uuid}.${filetype}`);
+        }
     },
 });
 const upload = multer({ storage: storage });
